fix: make sticky-bar handle animation
